Trying to Fix freedns sometimes failing:
Still under test:

It seems that there have been two issues that we have been dealing with lastly:
1) No prompt for entering API_KEY.
2) Problems with freedns registration.
Both have started lately in the last release, and don't happen all the time.
I think that they might be related, here is why.
I think that for some reason, asking for 10 reads is for some reason not enough to take us behind the pipelined input, so we might need to increase this. (for example to 1000)
This might also affect one running freedns, as the freedns is asking for email, and won't work without it.

tzachi-dar @tzachi-dar 11:17
The freedns might also fail because the network data is still not propagated. We are telling people to run things again. So, I'm saying, why don't we run it again for them.
That is after a failure in freedns, try to wait a minute and run it again.